flask/requirements.txt
numpy
tensorflow
opencv-python
flask
Pillow
flask-cors